Refactor sub-directories - securestore, direct_access_devicekey, conf
