+++ storage/kvstore/kv_config/mbed_lib.json
